Moved algorithm json download also to page where algorithms can be viewed only
