+++ vantage6-ui/src/app/components/algorithm/display-algorithm/display-algorithm.component.html
@@ -180,6 +180,18 @@
     </mat-accordion>
   </mat-card-content>
 </mat-card>
+<mat-card>
+  <mat-card-header>
+    <mat-card-title>{{ "algorithm-read.card-download-json.title" | translate }}</mat-card-title>
+  </mat-card-header>
+  <mat-card-content>
+    <p>{{ "algorithm-read.card-download-json.description" | translate }}</p>
+    <button mat-raised-button color="primary" (click)="downloadAlgorithmJson()">
+      <mat-icon>file_download</mat-icon>
+      <span>{{ "algorithm-read.card-download-json.download-button" | translate }}</span>
+    </button>
+  </mat-card-content>
+</mat-card>
 <ng-template #noFunctions>
   <p>{{ "algorithm-read.card-functions.no-functions" | translate }}</p>
 </ng-template>
